+++ src/index.jsx
@@ -1,10 +1,17 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './components';
-import * as serviceWorker from './serviceWorker';
+import React from 'react'
+import ReactDOM from 'react-dom'
+import { Provider } from 'redux-starter-kit/redux' 
+import './index.css'
+import App from './components'
+import * as serviceWorker from './serviceWorker'
+import store from './store'
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render( 
+    <Provider store={ store }>
+        <App />
+    </Provider>,
+    document.getElementById('root')
+)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
